// app/lib/models/persistence/quote.ts
export default interface Quote {
    asset_type: string
    asset_main_type: string
    cusip: string
    asset_sub_type: string
    symbol: string
    description: string
    bid_price: number
    bid_size: number
    bid_id: string
    ask_price: number
    ask_size: number
    ask_id: string
    last_price: number
    last_size: number
    last_id: string
    open_price: number
    high_price: number
    low_price: number
    bid_tick: string
    close_price: number
    net_change: number
    total_volume: number
    quote_time_in_long: bigint
    trade_time_in_long: bigint
    mark: number
    exchange: string
    exchange_name: string
    marginable: boolean
    shortable: boolean
    volatility: number
    digits: number
    fifty_two_wk_high: number
    fifty_two_wk_low: number
    nav: number
    pe_ratio: number
    div_amount: number
    div_yield: number
    div_date: string
    security_status: string
    regular_market_last_price: number
    regular_market_last_size: number
    regular_market_net_change: number
    regular_market_trade_time_in_long: number
    net_percent_change_in_double: number
    mark_change_in_double: number
    mark_percent_change_in_double: number
    regular_market_percent_change_in_double: number
    delayed: boolean
}